refactoring and folders managering
