+++ src/pages/Home/Pagination/Pagination.tsx
@@ -1,6 +1,6 @@
 import styles from "./Pagination.module.scss";
-import React, { useEffect, useState } from "react";
-import ReactDOM from "react-dom";
+import React from "react";
+
 import ReactPaginate from "react-paginate";
 
 type PaginationProps = {
