+++ src/pages/NotFound/NotFound.tsx
@@ -1,4 +1,3 @@
-import styles from "./NotFound.module.css";
 import NotFoundBlockInfo from "../../components/NotFoundBlock/NotFoundBlock";
 import React from "react";
