+++ src/redux/cart/selectors.ts
@@ -1,0 +1,3 @@
+import { RootState } from "../store";
+
+export const selectCart = (state: RootState) => state.cart;
