+++ src/redux/category/selectors.ts
@@ -1,0 +1,3 @@
+import { RootState } from "../store";
+
+export const selectCategory = (state: RootState) => state.category;
