+++ src/redux/category/types.ts
@@ -1,0 +1,4 @@
+export interface CategoryState {
+  categories: string[];
+  currentCategoryId: number;
+}
