+++ src/redux/filter/selectors.ts
@@ -1,0 +1,5 @@
+import { RootState } from "../store";
+
+export const selectFilter = (state: RootState) => state.filter;
+export const selectFilterBySort = (state: RootState) =>
+  state.filter.currentSortList.sort;
