+++ src/redux/filter/types.ts
@@ -1,0 +1,10 @@
+export type SortList = {
+  name: string;
+  sort: "rating" | "price" | "title";
+};
+
+export interface FilterState {
+  sortList: SortList[];
+  currentSortList: SortList;
+  isOpened: boolean;
+}
