+++ src/redux/search/types.ts
@@ -1,0 +1,3 @@
+export interface SearchState {
+  currentValue: string;
+}
