+++ src/utils/calcTotalItems.ts
@@ -1,4 +1,5 @@
-import { CartItem } from "../redux/slices/CartSlice";
+import { CartItem } from "../redux/cart/types";
+
 export const calcTotalItems = (items: CartItem[]) => {
   return items.reduce((sum, obj) => obj.count + sum, 0);
 };
